requirements.txt
markdown-it-py
anthropic
python-dotenv
openai